Lines are being drawn but with unintended connection behavior
